move: 테스트케이스 패키지 이동
CarsTest, CarTest, CarPositionTest, CarNameTest -> domain 패키지 이동
