+++ src/test/java/racingcar/domain/CarNameTest.java
@@ -1,13 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.CarName;
-import racingcar.domain.CarPosition;
 
 public class CarNameTest {
