+++ src/test/java/racingcar/domain/CarPositionTest.java
@@ -1,10 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.CarPosition;
 
 public class CarPositionTest {
     @Test
