+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
@@ -6,10 +6,6 @@ import generator.NumberGenerator;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.CarName;
-import racingcar.domain.CarPosition;
-import racingcar.domain.Cars;
 
 public class CarsTest {
